backend/app/requirements.txt
fastapi==0.98.0
uvicorn
pyjwt
pydantic==1.10.7 
jsonschema
orjson
typer
pytest
httpx
pymongo==3.11
python-dotenv
dnspython
motor
pytest==7.1.2
pytest-mock==3.6.1
openai
aiohttp
pyjwt
bcrypt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0     # PostgreSQL
pymysql>=1.0.2            # MySQL
